Functional Build
This is the first complete build of the game. The game was not initially
performing correctly; the "human" element touching an "enemy" element
did not end the game. Found a few strange lines of code while
troubleshooting, and when compared against the book's source code, they
weren't there; commenting them out resolved the issue.