End of Chapter 3
This is the end of chapter 3. As expected the page is broken, However
the provided source code from the web site does. I'm going to compare
the source code to what I got out of the book and correct it. While I
still don't think I'm going to continue working through this book, I
didn't like leaving what was uploaded in a broken state.