More Testing
More work done with SQL databases, mostly with DataGridViews and
Listboxes. My goals were to become more acquainted with SQL in C#, but
also to work with it to revamp my old VB.NET project for Battletech
random access tables. My hope is to remake the project in C#, and use a
database to store all the unit info, rather than the older method, which
was very disorganized.