began adding negative traits
I stopped because I need to think about exactly how I want to bring in
negative traits into the program. either just by parsing in the number
with a '-' in front (I believe that javas in parse will understand that)
or if I want to have a boolean negative value.