Successful test of reading xml file
The skills xml file read in perfectly. Populating a linked list of
skills with all of the skills listed in the xml file. It took in their
names as well as the boolean value for weather or not they are
difficult. While doing this I realized that I will need to adjust the
way that the xml file for traits is set up so that it will be able to
read into the program correctly.