big update to skills
skills is now read in properly through the XMLReader class, Game data
now contains the LinkedList of the skills for the program to work from
and will contain the data for the other LinkedLists as they are
finalized. Finally, skills now have the military boolean, as there are
often stages in life paths where characters get free skills to any
“military” skills. Unfortunately what is “military” is left at GM
discretion so I just had to do my best to assume what constitutes
military.