Massive Update
Traits are now implemented and are read in from their xml file. To
allow for the different methods that traits increment their value. 2
boolean values were added to accommodate this. One if the trait
increments by odd numbers (1,3,5,etc) or if it does it by addition
rather than multiplication. Also added the incrementPoints methods in
skills and traits, which should be used in place of the standard setters